discussion_insight_extractor/requirements.txt
google-genai
python-dotenv
markdown
weasyprint
langchain